+++ packages/workspaces-ui-core/changelog.md
@@ -1,3 +1,8 @@
+1.7.3
+feat: added support for the API frame initialization
+1.7.2
+feat: added isSelected to control how the workspace is being opened
+feat: pinned tabs support and workspace icons
 1.7.1
 chore: bump due to dependencies update
 1.7.0